requirements.txt
gensim>=4.3.2
underthesea>=6.8.0
pandas>=2.2.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
pytest>=8.0.0
pyyaml>=6.0.0
